+++ ManagedCorDebug/Managed/ResultTypes/GetDebugInfoWithPaddingResult.cs
@@ -1,14 +1,16 @@
+using System;
+
 namespace ManagedCorDebug
 {
     public struct GetDebugInfoWithPaddingResult
     {
-        public long PIDD { get; }
+        public IntPtr PIDD { get; }
 
         public int PcData { get; }
 
         public byte[] Data { get; }
 
-        public GetDebugInfoWithPaddingResult(long pIDD, int pcData, byte[] data)
+        public GetDebugInfoWithPaddingResult(IntPtr pIDD, int pcData, byte[] data)
         {
             PIDD = pIDD;
             PcData = pcData;
